feat(): re-organized files, updated styles and added work section
