+++ src/components/about/style.scss
@@ -6,7 +6,7 @@
 	overflow-y: scroll;
 }
 
-.sub-section {
+.about .sub-section {
 	width: 100%;
 	flex-grow: 1;
 	gap: 15px;
