+++ src/components/modules/core/style.scss
@@ -1,0 +1,26 @@
+.container {
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	width: 100%;
+	margin: 0 auto;
+	box-sizing: border-box;
+
+	&.round {
+		border-radius: 8px;
+	}
+
+	&.column {
+		flex-direction: column;
+	}
+
+	> .side-left,
+	> .side-right {
+		height: 100%;
+	}
+}
+
+.side-left,
+.side-right {
+	width: 50%;
+}
