+++ src/components/modules/transitions/index.js
@@ -1,0 +1,2 @@
+/* eslint-disable import/prefer-default-export */
+export { default as FadeTransition } from "./fade-transition";
