+++ src/components/page.js
@@ -1,9 +1,40 @@
 import React from "react";
-import Page from "./landing/component";
+import { createBrowserRouter, RouterProvider } from "react-router-dom";
+import PageLayout from "./layout";
+import Home from "./home";
+import { About } from "./about";
+import Works from "./work";
 import "../styles/app.scss";
 
+const Contact = () => <div>Contact</div>;
+
+const router = createBrowserRouter([
+	{
+		path: "/",
+		element: <PageLayout />,
+		children: [
+			{
+				path: "/",
+				element: <Home />,
+			},
+			{
+				path: "/about",
+				element: <About />,
+			},
+			{
+				path: "/works",
+				element: <Works />,
+			},
+			{
+				path: "/contact",
+				element: <Contact />,
+			},
+		],
+	},
+]);
+
 const App = () => {
-	return <Page />;
+	return <RouterProvider router={router} />;
 };
 
 export default App;
