+++ src/styles/app.scss
@@ -41,3 +41,7 @@ h3 {
 	font-size: 24px;
 	font-weight: 600;
 }
+
+a {
+	text-decoration: none;
+}
